requirements.txt
pillow
os
streamlit